requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.25.2
scipy>=1.10.1
diffusers==0.20.2
transformers==4.33.3
huggingface_hub==0.16.4
controlnet-aux==0.0.5
opencv-python==4.8.0.76
Pillow==10.0.0
accelerate==0.21.0
mediapip
